feat: split up actions
