+++ internal/actions/gather.go
@@ -1,0 +1,29 @@
+package actions
+
+import (
+	"context"
+	"fmt"
+	"net/http"
+
+	"github.com/promiseofcake/artifactsmmo-engine/internal/models"
+)
+
+// Gather performs resource gathering at the current position for the given character
+func (r *Runner) Gather(ctx context.Context, character string) (*SkillResponse, error) {
+	resp, err := r.Client.ActionGatheringMyNameActionGatheringPostWithResponse(ctx, character)
+	if err != nil {
+		return nil, fmt.Errorf("failed to gather: %w", err)
+	}
+
+	if resp.StatusCode() != http.StatusOK {
+		return nil, fmt.Errorf("failed to gather: %s (%d)", resp.Body, resp.StatusCode())
+	}
+
+	return &SkillResponse{
+		SkillInfo: resp.JSON200.Data.Details,
+		Response: Response{
+			CharacterResponse: models.Character{CharacterSchema: resp.JSON200.Data.Character},
+			CooldownSchema:    resp.JSON200.Data.Cooldown,
+		},
+	}, nil
+}
